Revert "pass APIEnablement through apiserver chain"
This reverts commit 2f403b7ad18a179514f1de77e29f1a2549ef030a.
