+++ cmd/kube-apiserver/app/options/validation.go
@@ -18,10 +18,6 @@ package options
 
 import (
 	"fmt"
-
-	apiextensionsapiserver "k8s.io/apiextensions-apiserver/pkg/apiserver"
-	aggregatorapiserver "k8s.io/kube-aggregator/pkg/apiserver"
-	"k8s.io/kubernetes/pkg/api/legacyscheme"
 )
 
 // TODO: Longer term we should read this from some config store, rather than a flag.
@@ -79,9 +75,5 @@ func (options *ServerRunOptions) Validate() []error {
 	if options.MasterCount <= 0 {
 		errors = append(errors, fmt.Errorf("--apiserver-count should be a positive number, but value '%d' provided", options.MasterCount))
 	}
-	if errs := options.APIEnablement.Validate(legacyscheme.Registry, apiextensionsapiserver.Registry, aggregatorapiserver.Registry); len(errs) > 0 {
-		errors = append(errors, errs...)
-	}
-
 	return errors
 }
